requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
python-dotenv==1.0.0 